Split Encoder module
